Menu/UI port-worklist closeout: magic chain, shop buy list, card draw family
- spell_menu: retail sub-screen mapping 0x0E..0x11 (PORT FUN_801D8F10/
  801D9110/801D9280/801D9594): Ra-Seru caster gate (NoRaSeru buzz),
  spell_targets_group (stats +2 bit 0x20 group/single routing)
- shop: buy_list_confirm_route (PORT FUN_801DB21C - gold gate + item-kind
  routing 1->recipient 2->quantity); save-screen.md 0x1A/0x1B/0x1E rows
  re-lensed as shop screens (0x1B 'card-full/error' falsified; 0x1A
  'save-slot confirm' scan is the bag, not save blocks)
- pause_screens: PORT FUN_801D7FF8 (all-party apply, count-0 picker)
- engine-ui: slot_grid_quad_x + grid PORT FUN_801E06C0/801E0FD0 (slide
  base 0x15A+slot*64 -> 0x5A, 104px park fan-out reconciled),
  backdrop_dim_sprites (PORT FUN_801E02A4), save_ui_record_quad (PORT
  FUN_801E3FF0), card_message_rows (PORT FUN_801E0418, dead pulse
  documented), confirm_prompt_draws (PORT FUN_801D1DAC/801D1F10),
  status panel tag FUN_801D33D8, options window tag FUN_801DCEF0,
  options sub-screen wrapper consts (PORT FUN_801DD330)
- docs: save-screen sub-screen table filled (items/magic/equip/shop
  rows), card I/O + drain + ticker + grid port notes; field-menu equip
  sub-screen chain + submenu slot map; functions.md +19 rows
- worklist CSV: 801d7a5c VA_ALIASED (0971 occupant undumped; dumped body
  is fishing 972); identifications recorded for 8 not-yet-ported rows

Gates: fmt, clippy engine-core/engine-ui -D warnings, cargo test
engine-core (2181 passed w/ LEGAIA_DISC_BIN) + engine-ui, md-link +
doc-density staged checks.

+++ crates/engine-core/src/prize_exchange.rs
@@ -410,7 +410,7 @@ mod tests {
         assert_eq!(coins, 4000);
         assert_eq!(inv.get(&0x20), Some(&1));
         assert!(set.contains(&0x36));
-        s.rebuild(|g| set.contains(&(g as u16)));
+        s.rebuild(|g| set.contains(&g));
         assert_eq!(s.rows().count(), 2);
     }
